RI-969354 Вынес модели в отдельные модули
