--- app/src/main/java/ru/rollinitiative/domain/model/Actor.kt
@@ -1,7 +1,0 @@
-package ru.rollinitiative.domain.model
-
-abstract class Actor {
-    abstract val id: Long
-    abstract val name: String
-    abstract val initiative: Int
-}
