--- app/src/main/java/ru/rollinitiative/domain/model/ActorSimple.kt
@@ -1,9 +1,0 @@
-package ru.rollinitiative.domain.model
-
-import ru.rollinitiative.domain.model.character.CreatureType
-
-data class ActorSimple(
-    val id: Int,
-    val name: String,
-    val type: CreatureType,
-)
