--- app/src/main/java/ru/rollinitiative/domain/model/CombatantGroup.kt
@@ -1,8 +1,0 @@
-package ru.rollinitiative.domain.model
-
-data class CombatantGroup(
-    override val id: Long,
-    override val name: String,
-    override val initiative: Int,
-    val combatants: List<Combatant>
-): Actor()
