--- app/src/main/java/ru/rollinitiative/domain/model/Dice.kt
@@ -1,13 +1,0 @@
-package ru.rollinitiative.domain.model
-
-enum class Dice(val value: Int) {
-    D4(4),
-    D6(6),
-    D8(8),
-    D10(10),
-    D12(12),
-    D20(20),
-    D100(100),
-}
-
-fun Dice.roll(): Int = (1..value).random()
