--- app/src/main/java/ru/rollinitiative/domain/model/Event.kt
@@ -1,8 +1,0 @@
-package ru.rollinitiative.domain.model
-
-class Event(
-    override val id: Long,
-    override val name: String,
-    override val initiative: Int,
-    val description: String,
-) : Actor()
