--- app/src/main/java/ru/rollinitiative/domain/model/character/CharacterSimple.kt
@@ -1,14 +1,0 @@
-package ru.rollinitiative.domain.model.character
-
-data class CharacterSimple(
-    override val id: Int,
-    override val name: String,
-    override val size: Size,
-    override val type: CreatureType,
-    override val armorClass: Int,
-    override val comment: String,
-    val maxHp: Int,
-) : Character() {
-
-    override fun getHP() = maxHp
-}
