--- app/src/main/java/ru/rollinitiative/domain/model/character/Characteristic.kt
@@ -1,10 +1,0 @@
-package ru.rollinitiative.domain.model.character
-
-enum class Characteristic {
-    Strength,
-    Dexterity,
-    Constitution,
-    Intelligence,
-    Wisdom,
-    Charisma,
-}
