--- app/src/main/java/ru/rollinitiative/domain/model/character/Characteristics.kt
@@ -1,10 +1,0 @@
-package ru.rollinitiative.domain.model.character
-
-data class Characteristics(
-    val strength: Int,
-    val dexterity: Int,
-    val constitution: Int,
-    val intelligence: Int,
-    val wisdom: Int,
-    val charisma: Int,
-)
