--- app/src/main/java/ru/rollinitiative/domain/model/character/DamageType.kt
@@ -1,19 +1,0 @@
-package ru.rollinitiative.domain.model.character
-
-enum class DamageType {
-    Slashing,
-    Piercing,
-    Bludgeoning,
-
-    Fire,
-    Cold,
-    Force,
-    Lightning,
-    Acid,
-    Poison,
-
-    Necrotic,
-    Psychic,
-    Thunder,
-    Radiant,
-}
