--- app/src/main/java/ru/rollinitiative/domain/model/character/HitPoint.kt
@@ -1,16 +1,0 @@
-package ru.rollinitiative.domain.model.character
-
-import ru.rollinitiative.domain.model.Dice
-import ru.rollinitiative.domain.model.roll
-
-data class HitPoint(
-    val dice: Dice,
-    val diceCount: Int,
-    val additionalHP: Int
-) {
-
-    fun calculate(): Int = mutableListOf<Int>()
-        .apply { repeat(diceCount) { add(dice.roll()) } }
-        .sumOf { it }
-        .and(additionalHP)
-}
