--- app/src/main/java/ru/rollinitiative/domain/model/character/Race.kt
@@ -1,5 +1,0 @@
-package ru.rollinitiative.domain.model.character
-
-enum class Race {
-    Elf, Human, Gnome, Dwarf, Halfling, HalfOrk, Dragonborn, Custom
-}
