--- app/src/main/java/ru/rollinitiative/domain/model/character/Size.kt
@@ -1,5 +1,0 @@
-package ru.rollinitiative.domain.model.character
-
-enum class Size {
-    TINY, SMALL, MEDIUM, BIG, HUGE, GIANT
-}
