--- app/src/main/java/ru/rollinitiative/domain/model/character/Skill.kt
@@ -1,30 +1,0 @@
-package ru.rollinitiative.domain.model.character
-
-enum class Skill {
-    // Strength
-    Athletics,
-
-    // Dexterity
-    Acrobatics,
-    SleightOfHand,
-    Stealth,
-
-    // Intelligence
-    Arcana,
-    History,
-    Investigation,
-    Nature,
-    Religion,
-
-    // Wisdom
-    AnimalHandling,
-    Insight,
-    Medicine,
-    Perception,
-
-    // Charisma
-    Deception,
-    Intimidation,
-    Performance,
-    Persuasion
-}
