--- app/src/main/java/ru/rollinitiative/domain/model/character/Speed.kt
@@ -1,8 +1,0 @@
-package ru.rollinitiative.domain.model.character
-
-data class Speed(
-    val running: Int? = null,
-    val flying: Int? = null,
-    val swimming: Int? = null,
-    val digging: Int? = null,
-)
