+++ core/model/domain/build.gradle.kts
@@ -1,0 +1,4 @@
+plugins {
+    id("ru.rollinitiative.java.library.plugin")
+    id("ru.rollinitiative.serialization.plugin")
+}
