+++ core/model/domain/src/main/kotlin/ru/rollinitiative/core/model/domain/DamageType.kt
@@ -1,0 +1,19 @@
+package ru.rollinitiative.core.model.domain
+
+enum class DamageType {
+    Slashing,
+    Piercing,
+    Bludgeoning,
+
+    Fire,
+    Cold,
+    Force,
+    Lightning,
+    Acid,
+    Poison,
+
+    Necrotic,
+    Psychic,
+    Thunder,
+    Radiant,
+}
