+++ feature/character/detail/api-logic/src/main/kotlin/ru/rolliniative/feature/character/detail/api/logic/model/Character.kt
@@ -1,0 +1,12 @@
+package ru.rolliniative.feature.character.detail.api.logic.model
+
+abstract class Character {
+    abstract val id: Int
+    abstract val name: String
+    abstract val size: Size
+    abstract val type: CreatureType
+    abstract val armorClass: Int
+    abstract val comment: String?
+
+    abstract fun getHP(): Int
+}
