+++ feature/character/detail/api-logic/src/main/kotlin/ru/rolliniative/feature/character/detail/api/logic/model/CharacterSimple.kt
@@ -1,0 +1,14 @@
+package ru.rolliniative.feature.character.detail.api.logic.model
+
+data class CharacterSimple(
+    override val id: Int,
+    override val name: String,
+    override val size: Size,
+    override val type: CreatureType,
+    override val armorClass: Int,
+    override val comment: String,
+    val maxHp: Int,
+) : Character() {
+
+    override fun getHP() = maxHp
+}
