+++ feature/character/detail/api-logic/src/main/kotlin/ru/rolliniative/feature/character/detail/api/logic/model/Characteristic.kt
@@ -1,0 +1,10 @@
+package ru.rolliniative.feature.character.detail.api.logic.model
+
+enum class Characteristic {
+    Strength,
+    Dexterity,
+    Constitution,
+    Intelligence,
+    Wisdom,
+    Charisma,
+}
