+++ feature/character/detail/api-logic/src/main/kotlin/ru/rolliniative/feature/character/detail/api/logic/model/Characteristics.kt
@@ -1,0 +1,10 @@
+package ru.rolliniative.feature.character.detail.api.logic.model
+
+data class Characteristics(
+    val strength: Int,
+    val dexterity: Int,
+    val constitution: Int,
+    val intelligence: Int,
+    val wisdom: Int,
+    val charisma: Int,
+)
