+++ feature/character/detail/api-logic/src/main/kotlin/ru/rolliniative/feature/character/detail/api/logic/model/HitPoint.kt
@@ -1,0 +1,16 @@
+package ru.rolliniative.feature.character.detail.api.logic.model
+
+import ru.rollinitiative.core.model.domain.Dice
+import ru.rollinitiative.core.model.domain.roll
+
+data class HitPoint(
+    val dice: Dice,
+    val diceCount: Int,
+    val additionalHP: Int
+) {
+
+    fun calculate(): Int = mutableListOf<Int>()
+        .apply { repeat(diceCount) { add(dice.roll()) } }
+        .sumOf { it }
+        .and(additionalHP)
+}
