+++ feature/character/detail/api-logic/src/main/kotlin/ru/rolliniative/feature/character/detail/api/logic/model/Race.kt
@@ -1,0 +1,5 @@
+package ru.rolliniative.feature.character.detail.api.logic.model
+
+enum class Race {
+    Elf, Human, Gnome, Dwarf, Halfling, HalfOrk, Dragonborn, Custom
+}
