+++ feature/character/detail/api-logic/src/main/kotlin/ru/rolliniative/feature/character/detail/api/logic/model/Size.kt
@@ -1,0 +1,5 @@
+package ru.rolliniative.feature.character.detail.api.logic.model
+
+enum class Size {
+    TINY, SMALL, MEDIUM, BIG, HUGE, GIANT
+}
