+++ feature/character/detail/api-logic/src/main/kotlin/ru/rolliniative/feature/character/detail/api/logic/model/Skill.kt
@@ -1,0 +1,30 @@
+package ru.rolliniative.feature.character.detail.api.logic.model
+
+enum class Skill {
+    // Strength
+    Athletics,
+
+    // Dexterity
+    Acrobatics,
+    SleightOfHand,
+    Stealth,
+
+    // Intelligence
+    Arcana,
+    History,
+    Investigation,
+    Nature,
+    Religion,
+
+    // Wisdom
+    AnimalHandling,
+    Insight,
+    Medicine,
+    Perception,
+
+    // Charisma
+    Deception,
+    Intimidation,
+    Performance,
+    Persuasion
+}
