+++ feature/character/detail/api-logic/src/main/kotlin/ru/rolliniative/feature/character/detail/api/logic/model/Speed.kt
@@ -1,0 +1,8 @@
+package ru.rolliniative.feature.character.detail.api.logic.model
+
+data class Speed(
+    val running: Int? = null,
+    val flying: Int? = null,
+    val swimming: Int? = null,
+    val digging: Int? = null,
+)
