+++ feature/character/list/api-logic/src/main/kotlin/ru/rolliniative/feature/character/list/api/logic/model/Actor.kt
@@ -1,0 +1,7 @@
+package ru.rolliniative.feature.character.list.api.logic.model
+
+abstract class Actor {
+    abstract val id: Long
+    abstract val name: String
+    abstract val initiative: Int
+}
