+++ feature/character/list/api-logic/src/main/kotlin/ru/rolliniative/feature/character/list/api/logic/model/ActorSimple.kt
@@ -1,0 +1,7 @@
+package ru.rolliniative.feature.character.list.api.logic.model
+
+data class ActorSimple(
+    val id: Int,
+    val name: String,
+    val type: CreatureType,
+)
