+++ feature/character/list/api-logic/src/main/kotlin/ru/rolliniative/feature/character/list/api/logic/model/Combatant.kt
@@ -1,0 +1,15 @@
+package ru.rolliniative.feature.character.list.api.logic.model
+
+import ru.rollinitiative.core.model.domain.Condition
+
+data class Combatant(
+    override val id: Long,
+    override val initiative: Int,
+    val character: Character,
+    val currentHp: Int,
+    val tmpHp: Int,
+    val conditions: List<Condition>
+): Actor() {
+
+    override val name: String get() = character.name
+}
