+++ feature/character/list/api-logic/src/main/kotlin/ru/rolliniative/feature/character/list/api/logic/model/CombatantGroup.kt
@@ -1,0 +1,8 @@
+package ru.rolliniative.feature.character.list.api.logic.model
+
+data class CombatantGroup(
+    override val id: Long,
+    override val name: String,
+    override val initiative: Int,
+    val combatants: List<Combatant>
+): Actor()
