+++ feature/character/list/api-logic/src/main/kotlin/ru/rolliniative/feature/character/list/api/logic/model/Event.kt
@@ -1,0 +1,8 @@
+package ru.rolliniative.feature.character.list.api.logic.model
+
+class Event(
+    override val id: Long,
+    override val name: String,
+    override val initiative: Int,
+    val description: String,
+) : Actor()
